backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
psycopg[binary]==3.2.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
pydantic[email]==2.11.4
pillow==11.2.1
pytesseract==0.3.13
PyMuPDF==1.25.5
email-validator==2.2.0
openai==2.8.1
google-cloud-documentai==3.2.0
google-auth==2.36.0